# perf(dam): cache memmap handler