--- a/jina/types/arrays/memmap.py
+++ b/jina/types/arrays/memmap.py
@@ -64,14 +64,12 @@ class DocumentArrayMemmap(
         self._header_path = os.path.join(path, 'header.bin')
         self._body_path = os.path.join(path, 'body.bin')
         self._key_length = key_length
-        self._invalidated_mmap = False  # a boolean flag to say mmap is invalidated, set on the change of mmap
         self._last_mmap = None  # the last cached mmap object
         self._load_header_body()
 
     @property
     def _mmap(self) -> 'mmap':
-        if self._invalidated_mmap or self._last_mmap is None:
-            self._invalidated_mmap = False
+        if self._last_mmap is None:
             self._last_mmap = mmap.mmap(
                 self._body_fileno, length=0, prot=mmap.PROT_READ
             )
@@ -122,9 +120,7 @@ class DocumentArrayMemmap(
             self._start = tmp[-1][1] + tmp[-1][3]
             self._body.seek(self._start)
 
-        self._invalidated_mmap = (
-            True  #: always invalid mmap after calling this function
-        )
+        self._last_mmap = None
 
     def __len__(self):
         return len(self._header_map)
@@ -138,7 +134,7 @@ class DocumentArrayMemmap(
             self.append(d, flush=False)
         self._header.flush()
         self._body.flush()
-        self._invalidated_mmap = True
+        self._last_mmap = None
 
     def clear(self) -> None:
         """Clear the on-disk data of :class:`DocumentArrayMemmap`"""
@@ -174,7 +170,7 @@ class DocumentArrayMemmap(
         if flush:
             self._header.flush()
             self._body.flush()
-        self._invalidated_mmap = True
+        self._last_mmap = None
 
     def __getitem__(self, key: Union[int, str]) -> 'Document':
         if isinstance(key, str):
@@ -212,7 +208,7 @@ class DocumentArrayMemmap(
         self._header.seek(0, 2)
         self._header.flush()
         self._header_map.pop(str_key)
-        self._invalidated_mmap = True
+        self._last_mmap = None
 
     def _str2int_id(self, key: str) -> int:
         return self._header_map[key][0]
@@ -237,13 +233,13 @@ class DocumentArrayMemmap(
                 self.append(value)
                 self._header_map[self._int2str_id(key)] = self._header_map[value.id]
                 del self[value.id]
-                self._invalidated_mmap = True
+                self._last_mmap = None
             else:
                 raise IndexError(f'`key`={key} is out of range')
         elif isinstance(key, str):
             value.id = key
             self.append(value)
-            self._invalidated_mmap = True
+            self._last_mmap = None
         else:
             raise TypeError(f'`key` must be int or str, but receiving {key!r}')