--- a/tests/unit/types/arrays/test_memmap.py
+++ b/tests/unit/types/arrays/test_memmap.py
@@ -246,3 +246,18 @@ def test_memmap_physical_size(tmpdir):
     assert da.physical_size == 0
     da.append(Document())
     assert da.physical_size > 0
+
+
+def test_memmap_append(tmpdir):
+    da = DocumentArrayMemmap(tmpdir)
+    d0 = Document(text='hello')
+    da.append(d0)
+    assert da[0] == d0
+    d1 = Document(text='world')
+    da.append(d1)
+    assert da[1] == d1
+
+    da2 = DocumentArrayMemmap(tmpdir)
+    assert len(da2) == 2
+    assert da2[0] == d0
+    assert da2[1] == d1