inlining `mmap.mmap` instead of opening it everytime saves about 30% time

on this branch:
```
get embedding ...	get embedding takes 4 seconds (4.64s)
get embedding ...	get embedding takes 4 seconds (4.84s)
get embedding ...	get embedding takes 4 seconds (4.68s)
```

on master (2aee836):
```
get embedding ...	get embedding takes 7 seconds (7.41s)
get embedding ...	get embedding takes 6 seconds (6.66s)
get embedding ...	get embedding takes 6 seconds (6.27s)

```

```python
from jina.logging.profile import TimeContext
from jina.types.arrays.memmap import DocumentArrayMemmap
from tests import random_docs

dam1 = DocumentArrayMemmap('./dam2')
dam1.clear()
dam1.extend(random_docs(100000))

with TimeContext('get embedding'):
    a = dam1.get_attributes('embedding')
    assert len(a) == 100000
```